co-locate things
