+++ .editorconfig
@@ -1,7 +1,7 @@
 root = true
 
 [*]
-indent_style = tab
+indent_style = spaces
 indent_size = 2
 charset = utf-8
 trim_trailing_whitespace = true
