+++ docker-compose.yml
@@ -1,0 +1,6 @@
+version: "3"
+services:
+  redis:
+    image: "redis"
+    ports:
+      - 6379:6379
