+++ web/package.json
@@ -1,63 +1,63 @@
 {
-	"name": "web",
-	"version": "0.1.0",
-	"private": true,
-	"dependencies": {
-		"@fortawesome/fontawesome-svg-core": "^1.2.28",
-		"@fortawesome/free-brands-svg-icons": "^5.13.0",
-		"@fortawesome/pro-duotone-svg-icons": "^5.13.0",
-		"@fortawesome/pro-light-svg-icons": "^5.13.0",
-		"@fortawesome/pro-solid-svg-icons": "^5.13.0",
-		"@fortawesome/react-fontawesome": "^0.1.9",
-		"@fullhuman/postcss-purgecss": "^1.3.0",
-		"@testing-library/jest-dom": "^4.2.4",
-		"@testing-library/react": "^9.3.2",
-		"@testing-library/user-event": "^7.1.2",
-		"autoprefixer": "^9.7.3",
-		"classnames": "^2.2.6",
-		"downshift": "^5.4.5",
-		"framer-motion": "1.10.3",
-		"history": "5.0.0-beta.5",
-		"lodash": "4.17.15",
-		"polished": "^3.4.2",
-		"postcss-cli": "^6.1.3",
-		"query-string": "6.11.1",
-		"react": "^16.13.0",
-		"react-dom": "^16.13.0",
-		"react-query": "^1.2.9",
-		"react-router": "6.0.0-alpha.5",
-		"react-router-dom": "6.0.0-alpha.5",
-		"react-scripts": "3.4.0",
-		"react-table": "^7.2.2",
-		"recoil": "^0.0.8",
-		"socket.io-client": "2.3.0",
-		"tailwindcss": "^1.4.0",
-		"typeface-cabin": "^0.0.72",
-		"typeface-catamaran": "^0.0.72",
-		"typeface-londrina-solid": "^0.0.71"
-	},
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject",
-		"build:styles": "postcss tailwind.css -o src/styles.css",
-		"prebuild": "yarn build:styles",
-		"prestart": "yarn build:styles"
-	},
-	"eslintConfig": {
-		"extends": "react-app"
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	}
+  "name": "web",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/free-brands-svg-icons": "^5.13.0",
+    "@fortawesome/pro-duotone-svg-icons": "^5.13.0",
+    "@fortawesome/pro-light-svg-icons": "^5.13.0",
+    "@fortawesome/pro-solid-svg-icons": "^5.13.0",
+    "@fortawesome/react-fontawesome": "^0.1.9",
+    "@fullhuman/postcss-purgecss": "^1.3.0",
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.3.2",
+    "@testing-library/user-event": "^7.1.2",
+    "autoprefixer": "^9.7.3",
+    "classnames": "^2.2.6",
+    "downshift": "^5.4.5",
+    "framer-motion": "1.10.3",
+    "history": "5.0.0-beta.5",
+    "lodash": "4.17.15",
+    "polished": "^3.4.2",
+    "postcss-cli": "^6.1.3",
+    "query-string": "6.11.1",
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0",
+    "react-query": "^1.2.9",
+    "react-router": "6.0.0-beta.0",
+    "react-router-dom": "^6.0.0-beta.0",
+    "react-scripts": "3.4.0",
+    "react-table": "^7.2.2",
+    "recoil": "^0.0.8",
+    "socket.io-client": "2.3.0",
+    "tailwindcss": "^1.4.0",
+    "typeface-cabin": "^0.0.72",
+    "typeface-catamaran": "^0.0.72",
+    "typeface-londrina-solid": "^0.0.71"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "build:styles": "postcss tailwind.css -o src/styles.css",
+    "prebuild": "yarn build:styles",
+    "prestart": "yarn build:styles"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
