--- web/src/atoms/play-queue.js
@@ -1,7 +1,0 @@
-import { atom } from "recoil";
-
-export const playQueueAtom = atom({
-	key: "crowdQ.storage.playQueue",
-	default: [],
-	persistence_UNSTABLE: true,
-});
