--- web/src/atoms/player.js
@@ -1,11 +1,0 @@
-import { atom } from "recoil";
-
-export const isPlayingAtom = atom({
-	key: "crowdQ.isPlaying",
-	default: false,
-});
-
-export const currentPlaybackAtom = atom({
-	key: "crowdQ.currentPlayback",
-	default: null,
-});
