+++ web/src/hooks/use-login.js
@@ -1,44 +1,93 @@
-import { useState, useCallback, useMemo } from "react";
-import { useSetRecoilState } from "recoil";
-import {
-	userAccessTokenAtom,
-	userRefreshTokenAtom,
-} from "atoms/user-credentials";
+import { useState } from "react";
+import { atom, selector, useRecoilState, useSetRecoilState } from "recoil";
+import { stringify } from "query-string";
 
 const { REACT_APP_API_BASE_URL: API_BASE_URL } = process.env;
 
+export const clientAccessTokenQuery = selector({
+  key: "crowdQ.clientAccessTokenQuery",
+  get: async () => {
+    const response = await fetch(`${API_BASE_URL}/authorize`);
+    return await response.json();
+  },
+});
+
+export const clientAccessTokenState = atom({
+  key: "crowdQ.clientAccessToken",
+  default: null,
+});
+
+export const userAccessTokenAtom = atom({
+  key: "crowdQ.storage.userAccessToken",
+  default: null,
+  persistence_UNSTABLE: true,
+});
+
+export const userRefreshTokenAtom = atom({
+  key: "crowdQ.storage.userRefreshToken",
+  default: null,
+  persistence_UNSTABLE: true,
+});
+
 export function useLogin() {
-	const setUserAccessToken = useSetRecoilState(userAccessTokenAtom);
-	const setUserRefreshToken = useSetRecoilState(userRefreshTokenAtom);
-	const [status, setStatus] = useState("idle");
+  const setUserAccessToken = useSetRecoilState(userAccessTokenAtom);
+  const [userRefreshToken, setUserRefreshToken] = useRecoilState(
+    userRefreshTokenAtom,
+  );
+  const [status, setStatus] = useState("idle");
 
-	const login = useCallback(async (code, redirect_uri, grant_type) => {
-		setStatus("pending");
+  async function login(code, redirect_uri, grant_type) {
+    setStatus("pending");
 
-		try {
-			const response = await fetch(`${API_BASE_URL}/login`, {
-				method: "POST",
-				body: JSON.stringify({ code, redirect_uri, grant_type }),
-				headers: {
-					"Content-Type": "application/json",
-				},
-			});
+    try {
+      const response = await fetch(`${API_BASE_URL}/login`, {
+        method: "POST",
+        body: JSON.stringify({ code, redirect_uri, grant_type }),
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
 
-			if (response.error) {
-				throw response.error;
-			}
+      if (response.error) {
+        throw response.error;
+      }
 
-			const { access_token, refresh_token } = await response.json();
+      const { access_token, refresh_token } = await response.json();
 
-			if (response.ok && access_token && refresh_token) {
-				setUserAccessToken(access_token);
-				setUserRefreshToken(refresh_token);
-				setStatus("success");
-			}
-		} catch (err) {
-			setStatus("failure");
-		}
-	});
+      if (response.ok && access_token && refresh_token) {
+        setUserAccessToken(access_token);
+        setUserRefreshToken(refresh_token);
+        setStatus("success");
+      }
+    } catch (err) {
+      setStatus("failure");
+    }
+  }
 
-	return { login, status };
+  async function refresh() {
+    setStatus("pending");
+    try {
+      const response = await fetch(
+        `${API_BASE_URL}/refresh?${stringify({
+          refresh_token: userRefreshToken,
+        })}`,
+      );
+
+      if (response.error) {
+        throw response.error;
+      }
+
+      const { access_token, refresh_token } = await response.json();
+
+      if (response.ok && access_token && refresh_token) {
+        setUserAccessToken(access_token);
+        setUserRefreshToken(refresh_token);
+        setStatus("success");
+      }
+    } catch (err) {
+      setStatus("failure");
+    }
+  }
+
+  return { login, refresh, status };
 }
