+++ web/src/hooks/use-playlists.js
@@ -1,8 +1,8 @@
 import { useRecoilValue, useRecoilValueLoadable } from "recoil";
-import { userDetailsSelector } from "../atoms/user-details";
-import { userPlaylistsQuery } from "../atoms/playlists";
+import { userDetailsSelector } from "atoms/user-details";
+import { userPlaylistsQuery } from "atoms/playlists";
 
 export function usePlaylists() {
-	const userDetails = useRecoilValue(userDetailsSelector);
-	return useRecoilValueLoadable(userPlaylistsQuery(userDetails.id));
+  const userDetails = useRecoilValue(userDetailsSelector);
+  return useRecoilValueLoadable(userPlaylistsQuery(userDetails.id));
 }
