+++ web/src/hooks/use-queue.js
@@ -1,20 +1,24 @@
-import { useRecoilState } from "recoil";
-import { playQueueAtom } from "../atoms/play-queue";
+import { atom, useRecoilState } from "recoil";
+
+export const playQueueAtom = atom({
+  key: "crowdQ.playQueue",
+  default: [],
+});
 
 export function useQueue() {
-	const [queue, setQueue] = useRecoilState(playQueueAtom);
+  const [queue, setQueue] = useRecoilState(playQueueAtom);
 
-	function addToQueue(item) {
-		setQueue([...queue, item]);
-	}
+  function addToQueue(item) {
+    setQueue([...queue, item]);
+  }
 
-	function removeFromQueue(index) {
-		setQueue([...queue.slice(0, index), ...queue.slice(index + 1)]);
-	}
+  function removeFromQueue(index) {
+    setQueue([...queue.slice(0, index), ...queue.slice(index + 1)]);
+  }
 
-	function nextTrackInQueue() {
-		setQueue(queue.slice(1));
-	}
+  function nextTrackInQueue() {
+    setQueue(queue.slice(1));
+  }
 
-	return { queue, addToQueue, removeFromQueue, nextTrackInQueue };
+  return { queue, addToQueue, removeFromQueue, nextTrackInQueue };
 }
